fix format version upgrade

+++ iceberg-rust/src/catalog/commit.rs
@@ -420,7 +420,7 @@ pub fn apply_table_updates(
     for update in updates {
         match update {
             TableUpdate::UpgradeFormatVersion { format_version } => {
-                if u8::from(metadata.format_version) == format_version as u8 {
+                if i32::from(metadata.format_version) == format_version {
                     return Ok(());
                 }
                 unimplemented!("Table format upgrade");
@@ -520,7 +520,7 @@ pub fn apply_view_updates<T: Materialization + 'static>(
     for update in updates {
         match update {
             ViewUpdate::UpgradeFormatVersion { format_version } => {
-                if u8::from(metadata.format_version.clone()) == format_version as u8 {
+                if i32::from(metadata.format_version.clone()) == format_version {
                     return Ok(());
                 }
                 unimplemented!("Upgrade of format version");
